solve conflit
